refactor: Update import paths for convertBinaryToString utility
