+++ src/parseBinaryToArraySync.ts
@@ -1,7 +1,7 @@
 import type { CSVRecord, ParseBinaryOptions } from "./common/types.ts";
 import { commonParseErrorHandling } from "./commonParseErrorHandling.ts";
-import { convertBinaryToString } from "./convertBinaryToString.ts";
 import { parseStringToArraySync } from "./parseStringToArraySync.ts";
+import { convertBinaryToString } from "./utils/convertBinaryToString.ts";
 
 export function parseBinaryToArraySync<Header extends ReadonlyArray<string>>(
   binary: Uint8Array | ArrayBuffer,
