+++ src/utils/convertBinaryToString.spec.ts
